webscrapper/requirements.txt
requests
python-dotenv
beautifulsoup4
pytest
pytest-mock


